backend/requirements.txt
cssselect2==0.4.1
defusedxml==0.7.1
et-xmlfile==1.1.0
loguru==0.5.3
numpy==1.21.3
openpyxl==3.0.9
xlrd 
pandas==1.3.4
Pillow==8.4.0
pycparser==2.20
pydantic==1.8.2
python-barcode==0.13.1
python-dateutil==2.8.2
python-Levenshtein==0.12.0
pytz==2021.3
six==1.16.0
svgwrite==1.4.1
tinycss2==1.1.0
typing-extensions==3.10.0.2
webencodings==0.5.1
requests==2.26.0

Django>=3.2.8
psycopg2-binary>=2.9.1
djangorestframework>=3.12.4
Markdown>=3.3.4
django-filter>=21.1
drf-yasg>=1.20.0
ipython>=7.28.0
pyQiwiP2P>=1.0.1
celery>=5.1.2
redis>=3.5.3
PyAMQP==0.1.0.7
pandas==1.3.4
Pillow==8.4.0
openpyxl
django-modeltranslation
django-import-export
astral==2.2
gunicorn
pyyaml
django_filter

django-constance
django-picklefield

watchdog==2.1.6

Scrapy==2.5.1
geopy==2.0.0
envparse==0.2.0
extruct==0.10.0
dukpy==0.2.3